[TS] modify the type of segment index from long to int

+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/PartitionRequestClient.java
@@ -85,7 +85,7 @@ public interface PartitionRequestClient {
     /**
      * Notify the upstream remember the required segmentId.
      */
-    void notifyRequiredSegmentId(long segmentId, RemoteInputChannel inputChannel);
+    void notifyRequiredSegmentId(int segmentId, RemoteInputChannel inputChannel);
 
     /**
      * Cancels the partition request for the given remote input channel and removes this client from
