+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/tieredstore/downstream/common/SingleChannelTierClient.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 /** The interface of {@link SingleChannelTierClient} in Tiered Store. */
 public interface SingleChannelTierClient {
 
-    //boolean hasSegmentId(InputChannel inputChannel, long segmentId) throws IOException;
+    //boolean hasSegmentId(InputChannel inputChannel, int segmentId) throws IOException;
 
     Optional<InputChannel.BufferAndAvailability> getNextBuffer(
-            InputChannel inputChannel, long segmentId) throws IOException, InterruptedException;
+            InputChannel inputChannel, int segmentId) throws IOException, InterruptedException;
 
     void close() throws IOException;
 }
